Stop dependabot from handling updates, too noisy
There's no need for us to have a PR for each submodule update, way too noisy and fine
grained.

Instead the plan is to move to a ~monthly Github Actions job that does a
`git submodule foreach git pull` (or lightweight equivalent) and opens a single PR
with all the submodules updates in it.